Activate 'log out' button
Linked the logout button in the App layout to the 'destroy_user_session'
path, which logs the user out, destroys her session, and redirect to the
login page.

This closes #10 for now. There's more to do as the app progresses, such
as replace the dummy content in the off-canvas menu with real
dynamically generated links once those components are in place.