Implements Context Menu
Replaces the dummy data on the off-page side menu with a dynamically
generated nested list of the User's contexts. Each item in the menu
links to the `#by_context` action with the context id, except "All"
which links to the site root**, and and "Inbox" which links to
the #by_context action using the "Inbox" name.

The menu itself is generated using the helper method
`#nested_context_menu` in `tasks_view_helpers`.

TODO: Add :before rule to the stylesheet to prepend context names with
their icons.

Closes #13