Corrects bug on embedded Tasks Detail view.
The status value on the embedded task detail portion (the pop over
seen when clicking the pencil icon on the task list view) wasn't being
updated when the 'complete' status was toggled by clicking the checkbox.
(The update on the server was going through okay.) Updated toggle.js.erb
to include an update to text of the status div of the detail view.