Updates Tasks#by_project to render :index view
- Changed `@tasks_by_project` refs to `@task_groups`
- Changed implicit render of :by_project view to explicitly render :index view
- Removed :by_project view

Tested in browser: OK
Specs: All green
Featurs: All green

TODO:

Current code does not render a task group if it contains no tasks. This is
good when listing multiple tasks groups, but when viewing a single empty
group it results in a blank page with only a title (Context or Project).

Should probably check to see if we're rendering a single group and display
a 'no tasks' message.

refs #36