Amends Task List to display empty group, if only one group is selected
Current code did not render a task group if it contains no tasks. This is
good when listing multiple tasks groups, but when viewing a single empty
group it resulted in a blank page with only a title (Context or Project).

View now checks to see if we're rendering a single group and displays
the group with a 'no tasks' message.

Closes #36