Converts Context::ICONS constant from Hash to Array
The ICONS constant was set as hash with names as keys and the Foundation
icon font character name as value. This was because we were thinking
that we'd add character names to class attributes and let Foundation
assign the icons to the elements. Ultimately decided instead to use our
own CSS rules to assign the icons.

Changed the Hash to an Array containing just the friendly icon names.