Cleans up Cuke Feature files
Combined related scenarios into a single feature file.

Also put in placeholders for a couple more scenarios. Lots more to do
here, but I'd like to get all the behavior of the app specified here
(even if it's just in English) such that I could hand it off to someone
building a different front end (React, iOS, whatever) and they would be
able to recreate all the significant features of the app.