Fixes URL field bug in Task Detail view
The URL field should be hidden in the Task Detail view
when it's empty, but sometimes (not always) it's wasn't.

The problem was that I was using `if task.url ...` which
evalutes to true for an empty string.

Changed to `unless task.url.empty? ...` and everything is
working as expected.

----

fixes #43