Adds URL field to Cuke testing Tasks
The tasks created background step definition
'Fred has several tasks' did not contain any data
for the URL field. This was causing a error when
we tried to call `.empty` on nil objects.

Added url data to each of the Tasks (incl. a
few empty strings).

All Features now green.